reworking many rest services after testing them

+++ src/main/java/io/github/Hattinger04/course/model/solution/SolutionRepository.java
@@ -10,5 +10,5 @@ public interface SolutionRepository extends JpaRepository<Solution, Long>{
 	Solution findById(int id); 
 	// TODO: SQL not tested yet!
 	@Query(value = "SELECT * FROM SOLUTION s JOIN exercise e USING (exercise_id) where exercise_id=:exercise_id and e.name=:name", nativeQuery = true)
-	Solution findByExercise(@Param("exercise_id")int exercise_id, @Param("name")String name);
+	Solution findByExercise(@Param("exercise_id")long exercise_id, @Param("name")String name);
 }
